This commit adds a much better module creation interface with integrated Makefiles. See the module_template under the modules directory and the Makefile under the modules directory. for more information. It's really simple and will automatically recompile when the kernel is updated. Also, I've tested the kernel on real hardware again (my little Atom D525) and it worked well (with some small bug fixes).
With the new module interface, I have started on an IDE driver module, and with that I have
implemented PCI enumeration and PCI bus searching. Currently, the IDE module simply searches
and finds the IDE controller on the PCI bus (based on PCI class/subclass pair). More to come.
Haha