This commit updates mainly the linker script, even though other files are editted.
There was a problem where I was getting sparatic errors on a real machine depending
one if I added new lines to the code. I realized it was because my linker script
wasn't aligning the code as it should have been. The code is now page aligned, and
all is well.